Migrate Next to main Stepper component
