+++ assets/javascripts/kitten/components/steppers/next/stepper.js
@@ -1,0 +1,1 @@
+export { Stepper } from '../stepper'
